requirements.txt
# FastAPI + ASGI server
fastapi==0.115.2
uvicorn[standard]==0.32.0

# HTTP client
httpx==0.27.2

# Database + ORM + Migrations
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.13.2

# Security / JWT
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3

# Utilities
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.4.0
email-validator==2.1.0

# AI / LLM Providers (Sprint 3)
google-generativeai==0.8.3
openai==1.55.3
anthropic==0.39.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0

# Production server (gunicorn with uvicorn workers)
gunicorn==22.0.0
